backend/requirements.txt
asgiref
cffi
cryptography
Django
django-cors-headers
django-rest-knox
djangorestframework
pillow
pycparser
sqlparse
